Client

See merge request !23

+++ core/include/network/packet/PacketDeleteEntity.hh
@@ -13,13 +13,13 @@ namespace network {
      class PacketDeleteEntity : public APacket {
      public:
         virtual ~PacketDeleteEntity();
-        PacketDeleteEntity(uint32_t tick = 0,uint32_t eventId = 0, uint16_t entityId = 0);
+        PacketDeleteEntity(uint32_t tick = 0,uint32_t eventId = 0, uint32_t entityId = 0);
          void setTick(uint32_t tick);
          uint32_t getTick() const;
          void setEventId(uint32_t id);
          uint32_t getEventId() const;
-         void setEntityId(uint16_t entityId);
-         uint16_t getEntityId() const;
+         void setEntityId(uint32_t entityId);
+         uint32_t getEntityId() const;
      };
  }
 }
